update frontend for description/req changes
