+++ client/src/lib/utils.tsx
@@ -1,4 +1,5 @@
 import { groupBy } from 'lodash';
+import { Link } from 'react-router-dom';
 
 import { Course } from '../model/Course';
 import { Instructor } from '../model/Instructor';
@@ -149,6 +150,8 @@ export const capitalize = (s: string): string =>
 export const punctuate = (s: string): string =>
   s.charAt(s.length - 1) === '.' ? s : s + '.';
 
+const COURSE_CODE_REGEX = /^(([A-Z0-9]){4} [0-9]{3}(D1|D2|N1|N2|J1|J2|J3)?)$/;
+
 /**
  * Validates if a string matches the course code format.
  * Valid format: 4 alphanumeric chars + space + 3 digits + optional suffix
@@ -157,7 +160,7 @@ export const punctuate = (s: string): string =>
  * @returns {boolean} True if string is a valid course code
  */
 export const isValidCourseCode = (s: string): boolean =>
-  /^(([A-Z0-9]){4} [0-9]{3}(D1|D2|N1|N2|J1|J2|J3)?)$/.test(s);
+  COURSE_CODE_REGEX.test(s);
 
 /**
  * Inserts a delimiter between the subject and number portions of a course code.
@@ -261,4 +264,46 @@ export const timeSince = (
   const years = Math.floor(months / 12);
 
   return years === 1 ? '1 year ago' : `${years} years ago`;
+};
+
+export const replaceCourseLinks = (html: string): React.ReactNode[] => {
+  const doc = new DOMParser().parseFromString(
+    capitalize(punctuate(html.trim())),
+    'text/html'
+  );
+
+  return Array.from(doc.body.childNodes).map((node, index) => {
+    switch (node.nodeType) {
+      case Node.ELEMENT_NODE: {
+        const elem = node as HTMLElement;
+
+        if (node.nodeName === 'A') {
+          const href = elem.getAttribute('href');
+
+          if (!href) return elem.innerText;
+
+          if (!isValidCourseCode(elem.innerText))
+            return <a href={href}>{elem.innerText}</a>;
+
+          const courseCode = elem.innerText.replace(' ', '-');
+
+          return (
+            <Link
+              key={index}
+              to={`/course/${courseCode}`}
+              className='text-gray-800 hover:underline dark:text-gray-200'
+            >
+              {elem.innerText}
+            </Link>
+          );
+        }
+
+        return <span key={index}>{(node as HTMLElement).innerText}</span>;
+      }
+      case Node.TEXT_NODE:
+        return <span key={index}>{(node as Text).textContent}</span>;
+      case Node.COMMENT_NODE:
+        return null;
+    }
+  });
 };
